Add 'value()' and 'spline()' functions in formula parsing (in evalp.cpp only)

+++ src/eval.cpp
@@ -132,7 +132,7 @@ EQ_CL		// Clausen function
 //-----------------------------------------------------------------------------
 int mglFormula::Error=0;
 bool MGL_LOCAL_PURE mglCheck(char *str,int n);
-int MGL_LOCAL_PURE mglFindInText(const char *str, const char *lst);
+long MGL_LOCAL_PURE mglFindInText(const char *str, const char *lst);
 //-----------------------------------------------------------------------------
 #if MGL_HAVE_GSL
 MGL_NO_EXPORT gsl_rng *mgl_rng=0;	// NOTE: should be deleted by gsl_rng_free() but I don't know where :(
@@ -760,7 +760,7 @@ bool MGL_LOCAL_PURE mglCheck(char *str,int n)
 }
 //-----------------------------------------------------------------------------
 // Try to find one of symbols lst in the string str
-int MGL_LOCAL_PURE mglFindInText(const char *str, const char *lst)
+long MGL_LOCAL_PURE mglFindInText(const char *str, const char *lst)
 {
 	long l=0,r=0,len=strlen(str);
 	for(long i=len-1;i>=0;i--)
